:recycle: Change size to float for better precision

+++ displayable/IDisplayable.hpp
@@ -14,9 +14,9 @@ class IDisplayable {
 public:
     virtual ~IDisplayable() = default;
     [[nodiscard]] virtual const ICoordinate &getPosition() const = 0;
-    [[nodiscard]] virtual int getSize() const = 0;
+    [[nodiscard]] virtual float getSize() const = 0;
     virtual void setPosition(const ICoordinate &position) = 0;
     virtual void setPosition(std::unique_ptr<ICoordinate> position) = 0;
-    virtual void setSize(int size) = 0;
+    virtual void setSize(float size) = 0;
     [[nodiscard]] virtual char getReplacingChar() const = 0;
 };
